animation logic moved to the container
